Remove integration tests from prxSticky tests
- Export icon name colleciton to icons stories
- Cleanup eused imports from stories
