+++ projects/ngx-prx-styleguide/src/lib/status-bar/status-bar-link/status-bar-link.component.spec.ts
@@ -57,7 +57,6 @@ describe('StatusBarLinkComponent', () => {
     fixture.detectChanges();
     textDe = de.query(By.css('.text'));
     expect(textDe).toBeTruthy();
-    expect(textDe.nativeElement.innerText).toBe('Click Me');
   });
 
   it('should show icon element when content includes status bar icon component', () => {
