+++ projects/ngx-prx-styleguide/src/lib/status-bar/status-bar.module.ts
@@ -5,10 +5,8 @@ import { StatusBarComponent } from './status-bar.component';
 import { StatusBarLinkComponent } from './status-bar-link/status-bar-link.component';
 import { StatusBarIconComponent } from './status-bar-icon/status-bar-icon.component';
 import { StatusBarImageComponent } from './status-bar-image/status-bar-image.component';
-import { IconModule, iconNames, iconColorOptions } from '../icon/icon.module';
+import { IconModule } from '../icon/icon.module';
 import { ImageModule } from '../image/image.module';
-
-export {iconNames, iconColorOptions };
 
 @NgModule({
   declarations: [StatusBarTextComponent, StatusBarComponent, StatusBarLinkComponent, StatusBarIconComponent, StatusBarImageComponent],
